Use go fmt
Go provides a formatter that provides a standard way to indent code.

It doesn't allow any options which makes all Go code look the same.  I
may mangle it but the idea is that: "go fmt is everyone's favorite
because it is noones favorite."

Using go fmt settles any arguments about style and you'll find that it
makes all Go code look familiar.

I formatted all of the code in this directory with:

  $ go fmt ./...

The tripple dots (ellipsis?) is a special notation that the tools use to
indicate that it should recurse a directory.