readConfig() should return an error instead of exiting
You'll see a common theme in Go of passing errors up the stack when you
can.

readConfig() seems to be a good candidate for this style since I can
imagine a case where you may want to, for example, call readConfig()
more than once.  Or maybe retry.

So I think it is kind of un-Go-like to have readConfig() call os.Exit().

The error type is an interface and any type that implements a function
that matches the signature below adheres to it:

    func (t T) Error() string

Where T is some type.  So you can create your own error value or simply
use the one in the fmt package.  I typically use fmt.Errorf() which
returns a formatted error.