Keep happy path to the left.
Something I learned to do in C and was happy to find is common in Go
is the notion of keeping your 'happy path' to the left.

That is, if you are testing for an error it is idimatic to handle the
error right away by either returning, continuing, breaking or otherwise
short-curcuiting whatever logic/block you're in immediately.

So I find that if-else conditions are fairly rare when checking for
errors.  When I see something like below:

  if err != nil {
    // handle error
  } else {
    // handle non-error state
  }

I prefer to convert thta to:

  if err != nil {
    return err
  }

  // handle non-error state

In general I think this is good coding but you'll find that this is a
particularly common pattern in Go.  "else" sticks out to me as a result.

There is intentionally a bug left in here.  I'll fix this in the next
commit.