Use short declarator to fetch configuration item.
Short declarators are strange to people who are used to C and new to Go.
I'll have to point you to the Go spec on this one but the short answer
here is that it only requires you to have at least one new variable on
the left-hand-side.

So it is okay to say:

  rname, ok := re["name"]

Then later within the same scope:

  azimuth, ok := re["azimuth"]

In the first case, bot rname and ok will be declared as the apropriate
types (string and bool) and in the second case only azimuth will be
declared and ok will be effectively assigned.  ok must be the same type
so you can't, for example, use a quick declaration to change the type of
an already delcared variable.