Limit scope where possible
This is a personal preference.  Even though it seems useless, I try to
always limit the scope of symbols where possible.

Go's if conditions allow you to execute a short declration AND tests a
boolean expression like:

if <short-decl>; <bool-expr> {
  <block>
}

Where the results of the short-decl are local to the block.